requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
asyncio==4.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.13.0
dnspython==1.16.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
motor==3.7.1
multidict==6.6.4
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
propcache==0.3.2
pycparser==2.22
pymongo==4.14.1
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.4
rnet==2.4.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
structlog==25.4.0
tenacity==9.1.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1